requirements.txt
selenium==3.141.0
python-decouple==3.3
sentry-sdk==0.19.2
